+++ scripts/github-askpass.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+case "$1" in
+  *Username*) printf '%s\n' 'x-access-token' ;;
+  *Password*) printf '%s\n' "$GITHUB_TOKEN" ;;
+  *) exit 1 ;;
+esac